move _func_in_func location
